requirements.txt
ipython==3.2.0
numpy==1.9.2
Pillow==2.9.0
scipy==0.15.1
wheel==0.24.0
